Common: Rename operating_system.py to environment.py
This name is more extendable, consistent with utilities in the agent
