+++ monkey/tests/monkey_island/utils.py
@@ -1,6 +1,6 @@
 from pathlib import WindowsPath
 
-from common.utils.operating_system import is_windows_os
+from common.utils.environment import is_windows_os
 
 if is_windows_os():
     import win32api
