+++ monkey/tests/unit_tests/monkey_island/cc/repositories/test_local_storage_file_repository.py
@@ -7,7 +7,7 @@ import pytest
 from tests.monkey_island.utils import assert_linux_permissions, assert_windows_permissions
 from tests.utils import raise_
 
-from common.utils.operating_system import is_windows_os
+from common.utils.environment import is_windows_os
 from monkey_island.cc import repositories
 from monkey_island.cc.repositories import LocalStorageFileRepository
